Made 'content' a symlink
`quartz sync` creates a "backup" of the content directory by doing a `mv`
and then a `git pull`. This makes the previous `content` directory
invalid if you have a shell open at the old directory, resulting in
errors - if you're at the shell prompt and do something like `ls` you
get an error. If you're in vim you get an error. You end up having to
`cd` into the parent and then go back, and this is for any shell that
might have been in the content directory tree. A real pain.

Changing it to a symlink seems to resolve the issue. So now `git` is
showing all those prior files as deleted, therefore commiting this.